// src/ui/components/solutions/PopoverBase/PopoverBase.types.ts
export type PopoverMode = "accent" | "neutral" | "contrast" | "brand";
export type PopoverTrigger = "click" | "hover" | "none";
export type TipPosition =
    | "top-center"
    | "top-left"
    | "top-right"
    | "bottom-center"
    | "bottom-left"
    | "bottom-right"
    | "left-center"
    | "left-top"
    | "left-bottom"
    | "right-center"
    | "right-top"
    | "right-bottom";